tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animationevent_init_time.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - elapsedTimeArg in initAnimationEvent</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-events-">
   <meta name="flags" content="dom">
   <meta name="assert" content="elapsedTimeArg in initAnimationEvent specifies
   the amount of time, in seconds, the animation has been running at the time of
   initialization">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = undefined;
         div.addEventListener('animationstart', function(evt) {
            passed = (passed === undefined && evt.elapsedTime == 4);
         });
         setTimeout(function() {
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS' : 'FAIL');
         }, 500);
         try {
            var evt = document.createEvent('AnimationEvent');
            evt.initAnimationEvent('animationstart', true, false, 'Test', 4);
            div.dispatchEvent(evt);
         } catch(e) {
            passed = false;
         }
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
   <div></div>
</body>
</html>
